Create steps slider component
