+++ app/components/setup/styles.js
@@ -6,25 +6,5 @@ export default StyleSheet.create({
     flex: 1,
     backgroundColor: Theme.background.main,
     paddingTop: 88
-  },
-  indexBar: {
-    height: 44,
-    flexDirection: 'row',
-    alignItems: 'center',
-    justifyContent: 'center'
-  },
-  dot: {
-    backgroundColor: Theme.background.white,
-    width: 10,
-    height: 10,
-    borderRadius: 4,
-    marginLeft: 8,
-    marginRight: 8,
-    borderWidth: 2,
-    borderColor: Theme.borderColors.main
-  },
-  dotActive: {
-    backgroundColor: Theme.background.secondary,
-    borderColor: Theme.background.secondary
   }
 });
